Not present in 450 DSA sheet.
This is a medium level question, in which you can alter only String1 (source) in order to make it equal to String2 (target). 
You have to return minimum cost(s) of operations required for the conversion.
You can only perform these operations:
1. Ignore : Cost 0 
2. Add : Cost 1
3. Replace : Cost 1
4. Delete : Cost 1

As this can be done by DP, hence the folder.